requirements.txt
Django~=4.0.7
djangorestframework~=3.13.1